Route structure
1. Moved api calls in apiCalls file

2. Added Routes (to route users, -> routing api call to specific point example: /signup. /login)

3. Added controllers (controling api calls -> logic of login, sign up goes inside of it)
