+++ client/src/components/SignUp.js
@@ -15,10 +15,12 @@ const SignUp = () => {
   const [success, setSuccess] = useState(false);
 
   useEffect(() => {}, []);
+
   // regex for emiall and password
   const passRegex =
     /^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[#$@!%&*?])[A-Za-z\d#$@!%&*?]{8,30}$/;
   const emailRegex = /^[^\s@]+@[^\s@]+\.[^\s@]+$/;
+
   // handle submit
   const handleSubmit = async (e) => {
     e.preventDefault();
